raw_files_functions/requirements.txt
google-cloud==0.34.0
google-cloud-bigquery==2.4.0
google-cloud-storage==1.32.0
zlib-wrapper==0.1.3
oauth2client==4.1.3
xmltodict
pandas
